+++ src/components/pages/claim-page/download-await/styled-components.tsx
@@ -42,5 +42,5 @@ export const ListItem = styled.li`
 `
 
 export const Link = styled.a`
-  color: ${props => props.theme.linkColor}
+  color: ${props => props.theme.linkColor};
 `