+++ src/components/pages/claim-page/wallets-list/index.tsx
@@ -16,14 +16,16 @@ import CoinabseWalletIcon from 'images/coinbase-wallet.png'
 import BrowserWalletIcon from 'images/browser-wallet.png'
 import ZerionWalletIcon from 'images/zerion-wallet.png'
 import RainbowWalletIcon from 'images/rainbow-wallet.png'
+import ImtokenWalletIcon from 'images/imtoken-wallet.png'
 import WalletConnectIcon from 'images/walletconnect-wallet.png'
 import ENSIcon from 'images/ens-logo.png'
 import { useConnect, Connector } from 'wagmi'
-import { TDropStep } from 'types'
+import { TDropStep, TWalletName } from 'types'
 import { AdditionalNoteComponent } from 'linkdrop-ui'
 import {  OverlayScreen } from 'linkdrop-ui'
 import * as dropActions from 'data/store/reducers/drop/actions'
 import * as userAsyncActions from 'data/store/reducers/user/async-actions'
+import * as dropAsyncActions from 'data/store/reducers/drop/async-actions'
 import { Dispatch } from 'redux'
 import { DropActions } from 'data/store/reducers/drop/types'
 import { PopupContents } from './components'
@@ -42,10 +44,12 @@ const mapStateToProps = ({
 
 const mapDispatcherToProps = (dispatch: IAppDispatch & Dispatch<DropActions>) => {
   return {
-    setAddress: () => dispatch(
-      dropActions.setStep('set_address')
-    ),
+    setAddress: () => dispatch(dropActions.setStep('set_address')),
     setStep: (step: TDropStep) => dispatch(dropActions.setStep(step)),
+    deeplinkRedirect: (
+      deeplink: string,
+      walletId: TWalletName
+    ) => dispatch(dropAsyncActions.deeplinkRedirect(deeplink, walletId)),
     updateUserData: (
       address: string,
       chainId: number
@@ -63,6 +67,10 @@ const defineOptionsList = (
   connectors: Connector<any, any, any>[],
   wallet: string | null,
   downloadStarted: () => void,
+  deeplinkRedirect: (
+    deeplink: string,
+    walletId: TWalletName
+  ) => void,
   isManual: boolean,
   chainId: number
 ) => {
@@ -131,7 +139,9 @@ const defineOptionsList = (
   const metamaskDeeplink = getWalletDeeplink('metamask', system, window.location.href, chainId)
   const metamaskOption = (injectedOption && !injectedOptionIsBrave) || !metamaskDeeplink ? undefined : {
     title: 'Metamask',
-    href: metamaskDeeplink,
+    onClick: () => {
+      deeplinkRedirect(metamaskDeeplink, 'metamask')
+    },
     icon: <WalletIcon src={MetamaskIcon} />,
     recommended: wallet === 'metamask'
   }
@@ -139,7 +149,9 @@ const defineOptionsList = (
   const trustDeeplink = getWalletDeeplink('trust', system, window.location.href, chainId)
   const trustOption = (injectedOption && !injectedOptionIsBrave) || !trustDeeplink ? undefined : {
     title: 'Trust Wallet',
-    href: trustDeeplink,
+    onClick: () => {
+      deeplinkRedirect(trustDeeplink, 'trust')
+    },
     icon: <WalletIcon src={TrustWalletIcon} />,
     recommended: wallet === 'trust'
   }
@@ -147,7 +159,9 @@ const defineOptionsList = (
   const coinbaseDeeplink = getWalletDeeplink('coinbase', system, window.location.href, chainId)
   const coinbaseOption = (injectedOption && !injectedOptionIsBrave) || !coinbaseDeeplink ? undefined : {
     title: 'Coinbase Wallet',
-    href: coinbaseDeeplink,
+    onClick: () => {
+      deeplinkRedirect(coinbaseDeeplink, 'coinbase')
+    },
     icon: <WalletIcon src={CoinabseWalletIcon} />,
     recommended: wallet === 'coinbase_wallet'
   }
@@ -164,8 +178,20 @@ const defineOptionsList = (
   const rainbowDeeplink = getWalletDeeplink('rainbow', system, window.location.href, chainId)
   const rainbowOption = (injectedOption && !injectedOptionIsBrave) || !rainbowDeeplink ? undefined : {
     title: 'Rainbow',
-    href: rainbowDeeplink,
+    onClick: () => {
+      deeplinkRedirect(rainbowDeeplink, 'rainbow')
+    },
     icon: <WalletIcon src={RainbowWalletIcon} />,
+    recommended: wallet === 'rainbow'
+  }
+
+  const imwtokenDeeplink = getWalletDeeplink('imtoken', system, window.location.href, chainId)
+  const imtokenOption = (injectedOption && !injectedOptionIsBrave) || !imwtokenDeeplink ? undefined : {
+    title: 'ImToken',
+    onClick: () => {
+      deeplinkRedirect(imwtokenDeeplink, 'imtoken')
+    },
+    icon: <WalletIcon src={ImtokenWalletIcon} />,
     recommended: wallet === 'rainbow'
   }
 
@@ -176,6 +202,7 @@ const defineOptionsList = (
     zerionOption,
     walletConnectOption,
     ensOption,
+    imtokenOption,
     trustOption,
     rainbowOption
   ]
@@ -189,7 +216,8 @@ const WalletsList: FC<ReduxType> = ({
   wallet,
   chainId,
   isManual,
-  campaignId
+  campaignId,
+  deeplinkRedirect
 }) => {
   const { open } = useWeb3Modal()
   const { connect, connectors } = useConnect()
@@ -205,6 +233,7 @@ const WalletsList: FC<ReduxType> = ({
     connectors,
     wallet,
     () => setStep('download_await'),
+    (deeplink: string, walletId: TWalletName) => deeplinkRedirect(deeplink, walletId),
     isManual,
     chainId as number
   )